refactor: move screenshotHelper into rendering/ (#463)
screenshotHelper is the pygame screenshot backend — it is imported by
PygameRenderer.captureScreenshot — so it belongs under rendering/, not
src/screen/. Pure relocation: git mv the module and its test, and
update the one import (screen.screenshotHelper -> rendering
.screenshotHelper). No behavior change.

This removes pygame from src/screen/screenshotHelper.py, one step of the
#463 output-seam/lifecycle cleanup toward "pygame only in rendering/".
743 passing; black-clean; --selftest OK.

Part of #463.
