+++ CHANGELOG.md
@@ -8,6 +8,7 @@ logged in detail below.
 
 | Date | Commits | Summary |
 |------|---------|---------|
+| 2026-06-14 | 1+ | refactor: Move `screenshotHelper` into `rendering/` (#463) — it is the pygame screenshot backend (imported by `PygameRenderer.captureScreenshot`), so it belonged under `rendering/`, not `src/screen/`. Pure relocation + import update (`screen.screenshotHelper` → `rendering.screenshotHelper`); the matching test moves to `tests/rendering/`. Removes pygame from `src/screen/screenshotHelper.py` — one step of the #463 "pygame only in rendering/" cleanup. No behavior change; 743 passing |
 | 2026-06-14 | 1+ | test: Prove the frontend seam — `NullRenderer` + `NullInputSource` (Phase 5, closes #439, epic #433) — add headless, pygame-free implementations of the `Renderer` and `InputSource` interfaces (drawing no-ops; queries return sensible defaults; `NullInputSource` optionally replays scripted `InputEvent` frames) and a test that is the **executable proof** the abstraction holds: a subprocess poisons `pygame` in `sys.modules` and then drives the `Screen` base run loop one full frame through the Null frontend — if anything on that path imported pygame it would die with `ImportError` instead of printing the sentinel. Plus interface-conformance + one-frame-render tests. +6 tests; 743 passing |
 | 2026-06-14 | 1+ | refactor: Input seam — `worldScreen` game loop (Phase 4c, frontend-abstraction epic #433, #438) — convert the last and largest input consumer off pygame: `_processEvents` now reads `InputEvent`s from the injected `InputSource` (`pollEvents()`) and dispatches on `EventType` (the duplicate `WINDOWRESIZED`/`VIDEORESIZE` branches collapse into one `WINDOW_RESIZE`); `handleKeyDownEvent`'s lone `pygame.K_ESCAPE` becomes `KeyCode.ESCAPE`; `handleMouseWheelEvent` reads `event.scrollY`; and all ~13 `pygame.mouse.get_pos()`/`get_pressed()` calls go through `inputSource.getMousePosition()`/`getMouseButtons()`. With this, **no `src/screen/` file reads pygame input** — the remaining pygame in `worldScreen` is output (minimap/overlay `Surface`/`SRCALPHA`/`BLEND`, image load) and frame pacing (`time.Clock`), deferred to the output-seam phases. +5 headless dispatch tests (event routing, Esc→options, Esc-dismisses-help); 737 passing. ⚠️ Real-time loop (movement, mouse-drag HUD, minimap zoom) needs a maintainer smoke-run |
 | 2026-06-13 | 1+ | ci: Enforce Black formatting with a `black --check src tests` gate (closes #428) — adds a CI step (pinned `black==23.1.0` to match `format.sh` contributors) that fails the build on any formatting drift, so the cleanup in the preceding `chore` can't re-accumulate. Pairs the one-time sweep with the durable guard #428 asked for |
