+++ src/rendering/screenshotHelper.py
